Stop using backing bitmaps for the keyboards
I'm not convinced that using backing bitmaps is much of a performance gain for
redraws, since normally there shouldn't be things drawn on top of the keyboard.
Most of the changes to the keyboard view are actually done by modifying the
bitmap and repainting it.

This changes the onDraw logic to re-draw the keys directly on the supplied
Canvas. Hopefully this will help reduce the keyboard's rather excessive memory
consumption (cf. issue 64), the bitmaps are >1MB on a tablet or high-res phone
and get reallocated fairly frequently.

Need to watch out for performance regressions on odd platforms, though apparently the
Canvas supplied by the default View onDraw logic internally uses yet another backing
bitmap so this should not make a big difference. TODO: add option to get rid of that
bitmap as well.