Disable hardware acceleration
It's unclear if this is a net win for the keyboard, since hardware acceleration
apparently adds additional memory overhead and the keyboard doesn't really need
the extra speed (if any). It's doubtful if this provided much benefit in the
earlier version anyway, since apparently drawing to in-memory bitmaps isn't
accelerated, and modifying the bitmap requires re-uploading it to the GPU.

TODO: add an option to let the user select a rendering method at runtime.