Get rid of mModShift in LatinIME
Two places to track the same state == bad. Keep state in the Keyboard as it's supposed to be. This should fix shift-Arrow for multitouch and other cases.

Also clean up the 5-row portrait extended Fn map to add a shift key with status indicator, and prevent accidentally entering the unused SymbolsShifted mode.