Move popup layouts from LatinKeyboardBaseView to LatinKeyboardView
See issue 150.

Grumble. The original code is really confusing with its distinction between LatinKeyboardBaseView and the derived LatinKeyboardView - the former class is used for popups, the latter for the regular keyboard view. Neither should be confused with the Keyboard/LatinKeyboard classes which represent abstract key grids for use by the keyboard views, and the LatinIME and KeyboardSwitcher classes which tie them together.

Inexplicably, the code for adding long-press popups is in LatinKeyboardBaseView even though only the regular keyboard should be doing this. Android doesn't permit popup windows to be opened from popup windows, and due to unexpected default values apparently the base class decides to do so anyway.

This CL starts moving some of the code which should never be run for popup keyboards to LatinKeyboardView, though this is messy to untangle.